% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scHOTTER-package.R
\docType{package}
\name{scHOTTER-package}
\alias{scHOTTER}
\alias{scHOTTER-package}
\title{scHOTTER: scHOTTER: Parametric testing for spatially heterogeneous co-expression in single cell data}
\description{
scHOTTER is a package that implements kernel-based local correlation analysis for gene pairs and an analytic null for the sample variance of Fisher z correlations across kernels that comprises a parametric alternative to Ghazanfar et al.'s single-cell higher-order testing (scHOT) methodology (Ghazanfar et al., 2020). This pipeline offers improvements in computational efficiency and statistical power, supports both 2D spatial and 1D trajectory (1D) data, and provides a high-level pipeline with access to all intermediate results.
}
\author{
\strong{Maintainer}: Liam Skender \email{liamskender@protonmail.com}

Other contributors:
\itemize{
  \item Shila Ghazanfar [contributor]
}

}
\keyword{internal}
